move ethers stuff from protocol to client
