+++ packages/client/src/index-exports.ts
@@ -9,6 +9,8 @@ export * from './Subscriber'
 export * from './LoginEndpoints'
 export * from './StreamEndpoints'
 export { SearchStreamsOptions } from './StreamRegistry'
+import { getStorageNodeRegistryFromContract } from './getStorageNodeRegistryFromContract'
+export { getTrackerRegistryFromContract } from './getTrackerRegistryFromContract'
 import ConfigTest from './ConfigTest'
 import validateConfig from './ConfigBase'
