+++ packages/protocol/package.json
@@ -19,18 +19,15 @@
     "eslint": "eslint --cache --cache-location=node_modules/.cache/.eslintcache/ '*/**/*.{js,ts}'",
     "test": "jest",
     "test-unit": "jest test/unit --detectOpenHandles",
-    "test-integration": "jest test/integration --detectOpenHandles",
     "coverage": "jest --coverage",
     "docs": "typedoc --options typedoc.js"
   },
   "author": "Streamr",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@ethersproject/wallet": "^5.5.0",
     "@streamr/dev-config": "^1.0.0",
     "@types/heap": "^0.2.28",
     "@types/jest": "^27.0.2",
-    "@types/promise-memoize": "^1.2.1",
     "@types/sinon": "^10.0.4",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
@@ -43,19 +40,13 @@
     "streamr-test-utils": "^2.0.0",
     "ts-jest": "^27.0.5",
     "typedoc": "^0.22.5",
-    "typescript": "^4.5.2",
-    "web3-eth-accounts": "^1.6.0"
+    "typescript": "^4.5.2"
   },
   "dependencies": {
-    "@babel/runtime-corejs3": "^7.16.3",
-    "@ethersproject/contracts": "^5.5.0",
     "@ethersproject/hdnode": "^5.4.0",
-    "@ethersproject/providers": "^5.5.0",
     "@types/debug": "^4.1.7",
     "debug": "^4.3.2",
-    "events": "^3.3.0",
     "heap": "^0.2.6",
-    "promise-memoize": "^1.2.1",
     "secp256k1": "^4.0.2",
     "sha3": "^2.1.4",
     "strict-event-emitter-types": "^2.0.0"
