+++ packages/protocol/src/utils/StorageNodeRegistry.ts
@@ -1,11 +1,3 @@
-import { Contract } from '@ethersproject/contracts'
-import { JsonRpcProvider } from '@ethersproject/providers'
-
-type ProviderConnectionInfo = ConstructorParameters<typeof JsonRpcProvider>[0]
-
-import * as storageNodeRegistryConfig from '../../contracts/NodeRegistry.json'
-
-// storageNodeAddress => HTTP address
 export type StorageNodeInfo = {
     address: string
     url: string
@@ -31,40 +23,7 @@ export class StorageNodeRegistry {
         return this.records
     }
 }
-
-async function fetchStorageNodes(contractAddress: string, jsonRpcProvider: ProviderConnectionInfo): Promise<StorageNodeInfo[]> {
-    const provider = new JsonRpcProvider(jsonRpcProvider)
-    // check that provider is connected and has some valid blockNumber
-    await provider.getBlockNumber()
-
-    const contract = new Contract(contractAddress, storageNodeRegistryConfig.abi, provider)
-    // check that contract is connected
-    await contract.addressPromise
-
-    if (typeof contract.getNodes !== 'function') {
-        throw Error(`getNodes function is not defined in smart contract (${contractAddress})`)
-    }
-
-    const result = await contract.getNodes()
-    return result.map((node: any) => {
-        return {
-            address: node.nodeAddress,
-            url: JSON.parse(node.metadata).http
-        }
-    })
-}
-
 export function createStorageNodeRegistry(servers: StorageNodeInfo[]): StorageNodeRegistry {
     return new StorageNodeRegistry(servers)
 }
 
-export async function getStorageNodeRegistryFromContract({
-    contractAddress,
-    jsonRpcProvider
-}: {
-    contractAddress: string,
-    jsonRpcProvider: ProviderConnectionInfo
-}): Promise<StorageNodeRegistry> {
-    const storageNodes = await fetchStorageNodes(contractAddress, jsonRpcProvider)
-    return createStorageNodeRegistry(storageNodes)
-}
