+++ packages/protocol/src/utils/TrackerRegistry.ts
@@ -1,12 +1,5 @@
-import { Contract } from '@ethersproject/contracts'
-import { JsonRpcProvider } from '@ethersproject/providers'
-
 import { keyToArrayIndex } from './HashUtil'
-
-import * as trackerRegistryConfig from '../../contracts/TrackerRegistry.json'
 import { StreamPartID } from "./StreamPartID"
-
-type ProviderConnectionInfo = ConstructorParameters<typeof JsonRpcProvider>[0]
 
 export type SmartContractRecord = {
     id: string
@@ -33,48 +26,6 @@ export class TrackerRegistry<T extends TrackerInfo> {
     }
 }
 
-async function fetchTrackers(contractAddress: string, jsonRpcProvider: ProviderConnectionInfo) {
-    const provider = new JsonRpcProvider(jsonRpcProvider)
-    // check that provider is connected and has some valid blockNumber
-    await provider.getBlockNumber()
-
-    const contract = new Contract(contractAddress, trackerRegistryConfig.abi, provider)
-    // check that contract is connected
-    await contract.addressPromise
-
-    if (typeof contract.getNodes !== 'function') {
-        throw Error(`getNodes function is not defined in smart contract (${contractAddress})`)
-    }
-
-    return await contract.getNodes()
-}
-
 export function createTrackerRegistry<T extends TrackerInfo>(servers: T[]): TrackerRegistry<T> {
     return new TrackerRegistry(servers)
-}
-
-export async function getTrackerRegistryFromContract({
-    contractAddress,
-    jsonRpcProvider
-}: {
-    contractAddress: string,
-    jsonRpcProvider: ProviderConnectionInfo
-}): Promise<TrackerRegistry<SmartContractRecord>> {
-    const trackers = await fetchTrackers(contractAddress, jsonRpcProvider)
-    const records: SmartContractRecord[] = []
-    for (let i = 0; i < trackers.length; ++i) {
-        const { metadata, url, nodeAddress } = trackers[i]
-        try {
-            // The field is tracker.metadata in newer contracts and tracker.url in old contracts.
-            // It's safe to clean up tracker.url when no such contract is used anymore.
-            const urls = JSON.parse(metadata || url)
-            records.push({
-                id: nodeAddress,
-                ...urls
-            })
-        } catch (e) {
-            throw new Error(`Element trackers[${i}] not parsable as object: ${trackers[i]}`)
-        }
-    }
-    return createTrackerRegistry(records)
 }
